requirements.txt
streamlit>=1.35.0
anthropic>=0.25.0
voyageai>=0.2.3
supabase>=2.3.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
